+++ CHANGELOG.md
@@ -21,6 +21,7 @@ incremented for features.
 
 * lang: Add `ErrorCode::AccountNotInitialized` error to separate the situation when the account has the wrong owner from when it does not exist (#[1024](https://github.com/project-serum/anchor/pull/1024))
 * lang: Called instructions now log their name by default. This can be turned off with the `no-log-ix-name` flag ([#1057](https://github.com/project-serum/anchor/pull/1057))
+* lang: Add `ProgramData` AccountInfo wrapper that checks that given AccountInfo is owned by the upgradable loader and that its deserialization is the `ProgramData` variant of [UpgradeableLoaderState](https://docs.rs/solana-program/latest/solana_program/bpf_loader_upgradeable/enum.UpgradeableLoaderState.html) ([#1095](https://github.com/project-serum/anchor/pull/1095))
 
 ## [0.18.2] - 2021-11-14